@@ -1,4 +1,4 @@
-package ru.skypro.homework.exception;
+package ru.skypro.homework.exceptions;
 
 public class WrongPasswordException extends RuntimeException {
     public WrongPasswordException(String message) {
